__init__ cleanup to utils
